+++ extensions/identity-hub/src/main/java/org/eclipse/dataspaceconnector/identityhub/models/Status.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.identityhub.dtos;
+package org.eclipse.dataspaceconnector.identityhub.models;
 
 import io.swagger.v3.oas.annotations.media.Schema;
