+++ extensions/identity-hub/src/main/java/org/eclipse/dataspaceconnector/identityhub/models/WebNodeInterfaceMethod.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.identityhub.dtos;
+package org.eclipse.dataspaceconnector.identityhub.models;
 
 import java.util.stream.Stream;
